% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio.R
\name{bcbio_outputs}
\alias{bcbio_outputs}
\title{Get bcbio final output files}
\usage{
bcbio_outputs(final)
}
\arguments{
\item{final}{Path to \code{final} bcbio directory.}
}
\value{
A tibble with the following columns:
\itemize{
\item fpath: path to file
\item ftype: file type. Can be one of:
\enumerate{
\item ensembl-batch
\item mutect2-batch
\item strelka2-batch
\item vardict-batch
\item ensemble-germ
\item gatk-germ
\item strelka2-germ
\item vardict-germ
}
}
}
\description{
Generates a tibble containing absolute path to bcbio final files.
}
\examples{
\dontrun{
final <- "path/to/bcbio/final"
bcbio_outputs(final)
}
}
