% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcr-package.R
\docType{package}
\name{bcr-package}
\alias{bcr}
\alias{bcr-package}
\title{bcr}
\description{
Helper functions for bcbio run comparisons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umccr/bcr}
  \item Report bugs at \url{https://github.com/umccr/bcr/issues}
}

}
\author{
\strong{Maintainer}: Peter Diakumis \email{peterdiakumis@gmail.com} (0000-0002-7502-7545)

}
