% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{check_comparison_input}
\alias{check_comparison_input}
\title{Check Comparison Input File}
\usage{
check_comparison_input(x)
}
\arguments{
\item{x}{Path to tab-separated comparison input file.
Must contain the following five fields per file (NOTE: do not include column names):
sample name, variant type (SNV, SV or CNV), file label, run1 & run2 file paths (can be S3 paths).}
}
\value{
If the file has issues, an error message. If there
are no issues, simply returns file path.
}
\description{
Checks that the input file is suitable for running 'woof compare'.
}
\examples{
\dontrun{
x <- "path/to/cromwell_inputs.tsv"
check_comparison_input(x)
}
}
