% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcWrapper.R
\name{chkGnuBc}
\alias{chkGnuBc}
\title{Checks if GNU bc is availiable in the system}
\usage{
chkGnuBc(verbose = FALSE)
}
\arguments{
\item{verbose}{A \link{logical} which controls printing the GNU bc version.}
}
\description{
Checks if GNU bc is availiable in the system
}

