% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_purple_gene_files}
\alias{compare_purple_gene_files}
\title{Compare two PURPLE gene CNV files}
\usage{
compare_purple_gene_files(cnv1, cnv2, out_cn_diff, out_coord_diff,
  threshold = 0.1)
}
\arguments{
\item{cnv1}{Path to first PURPLE \code{gene.cnv} (or \code{cnv.gene.tsv}) file.}

\item{cnv2}{Path to second PURPLE \code{gene.cnv} (or \code{cnv.gene.tsv}) file ('TRUTHSET').}

\item{out_cn_diff}{Path to write genes with copy number differences between
the two runs - needs to be a writable directory.}

\item{out_coord_diff}{Path to write genes with different coordinates between
the two runs - needs to be a writable directory.}

\item{threshold}{Numeric. Copy number differences smaller than this value are not reported.}
}
\value{
Writes results from the comparison to \code{out_cn_diff} and \code{out_coord_diff}.
}
\description{
Compares two PURPLE \code{gene.cnv} (or \code{cnv.gene.tsv}) files.
}
\examples{

cnv1 <- system.file("extdata", "cnv/sample_A.purple.gene.cnv", package = "woofr")
cnv2 <- system.file("extdata", "cnv/sample_B.purple.cnv.gene.tsv", package = "woofr")
out1 <- tempfile()
out2 <- tempfile()
compare_purple_gene_files(cnv1, cnv2, out1, out2, threshold = 0.1)

}
