% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultar_bcrp_api.R
\name{conseguir_metadatos}
\alias{conseguir_metadatos}
\title{Extrae todos los códigos de serie disponibles dentro de BCRPData}
\usage{
conseguir_metadatos(x)
}
\arguments{
\item{x}{No requerido.}
}
\value{
Un data frame.
}
\description{
\code{conseguir_metadatos()} consigue el listado de códigos más actualizado directamente de BCRPData.
}
\examples{
todos_los_codigos <- conseguir_metadatos()

# A partir de este data.frame podemos explorar la cantidad de datos que tenemos disponible.

subset(todos_los_codigos,grepl("PBI",`Categoría de serie`))
}
