% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultar_bcrp_api.R
\name{consultar_bcrp_api}
\alias{consultar_bcrp_api}
\title{Consulta la API del BCR}
\usage{
consultar_bcrp_api(codigos, desde = NULL, hasta = NULL, idioma = "esp", ...)
}
\arguments{
\item{codigos}{String. Vector conteniendo los códigos de serie respectivos. Los códigos deben tener una misma frecuencia, de otro modo se elegirá la frecuencia del primer elemento del vector y se mostrarán las series que correspondan a esa frecuencia.}

\item{desde}{Opcional. Entero. Periodo inicial.}

\item{hasta}{Opcional. Entero. Periodo final.}

\item{idioma}{Idioma en el que se retornará el resultado. Por defecto igual a "esp" (Español), aunque es posible también "ing" (Inglés).}

\item{...}{Opcional. Información adicional a enviar junto con la solicitud.}
}
\value{
Un data frame.
}
\description{
\code{consultar_bcrp_api()} consulta y recibe los datos obtenidos del \href{https://estadisticas.bcrp.gob.pe/estadisticas/series/ayuda/api}{API para desarrolladores}.
}
\examples{
consultar_bcrp_api(codigos = "PN38705PM", desde = 2013, hasta = 2024)

# Usa ... para enviar información adicional dentro de la cabecera de la solicitud, con propósitos de identificación y transparencia.
consultar_bcrp_api(codigos = "PN01770AM", desde = 2020, hasta = 2024, `User-Agent` = "Hola desde R.")
}
