% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBase.R
\name{convertBase}
\alias{convertBase}
\title{Convert base of a numeric (vectorized)}
\usage{
convertBase(x, inputBase, outputBase)
}
\arguments{
\item{x}{A numeric to convert}

\item{inputBase}{A character among \code{c('2', '10', '16')}}

\item{outputBase}{A character among \code{c('2', '10', '16')}}
}
\description{
Convert base of a numeric (vectorized)
}

