% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_compound_operator.R
\name{cpp_compound}
\alias{cpp_compound}
\alias{\%+=\%}
\alias{\%*=\%}
\alias{\%-=\%}
\alias{\%/=\%}
\title{C++ style compound operators}
\usage{
left \%+=\% right

left \%*=\% right

left \%-=\% right

left \%/=\% right
}
\arguments{
\item{left}{left operand}

\item{right}{right operand}
}
\value{
the result
}
\description{
C++ style compound operators
}
\examples{
a <- 1
b <- 2
a \%+=\% 1
a
a \%+=\% b
a
b
a <- 3
b <- 2
a \%*=\% 5
a
a \%*=\% b
a
b
a <- 3
b <- 2
a \%-=\% 1
a
a \%-=\% b
a
b
a <- 3
b <- 2
a \%/=\% 3
a
a \%/=\% b
a
b
}
