% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_increment.R
\name{cpp_increment}
\alias{cpp_increment}
\alias{inc}
\alias{dec}
\title{C++ style increment / decrement operators}
\usage{
inc(var)

dec(var)
}
\arguments{
\item{left}{left operand}
}
\value{
the result
}
\description{
C++ style increment / decrement operators
}
\examples{
a <- 1
inc(a)
a
a <- 1
dec(a)
a
}
