% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{guess_file_type}
\alias{guess_file_type}
\title{Guess Type of File}
\usage{
guess_file_type(file)
}
\arguments{
\item{file}{File path.}
}
\value{
A character string: FASTQ, BAM, VCF or OTHER.
}
\description{
Guesses type of file.
}
