% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrd.R
\name{hrd_cmp}
\alias{hrd_cmp}
\title{HRDetect/CHORD Comparison}
\usage{
hrd_cmp(tool, f1, f2, out)
}
\arguments{
\item{tool}{hrdetect or chord?}

\item{f1}{Path to file1.}

\item{f2}{Path to file2.}

\item{out}{Path to output file with results.}
}
\description{
HRDetect/CHORD Comparison
}
