% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lf}
\alias{lf}
\title{List Files in knitr kable}
\usage{
lf(..., row.names = TRUE)
}
\arguments{
\item{...}{Passed to \code{list.files}}

\item{row.names}{Passed to \code{knitr::kable}}
}
\value{
A \code{knitr::kable}
}
\description{
Lists the files in the given directory in a single-column
\code{knitr::kable}.
}
\examples{
\dontrun{
lf("/path/to/foo")
lf("/path/to/foo", row.names = FALSE)
}
}
