% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{list_s3_objects}
\alias{list_s3_objects}
\title{List AWS S3 Objects}
\usage{
list_s3_objects(b, recursive = FALSE)
}
\arguments{
\item{b}{S3 path.}

\item{recursive}{Recursively list objects within given path.}
}
\value{
A tibble with file basename, complete S3 file path, and file type.
}
\description{
Lists objects inside S3 bucket.
}
\examples{
\dontrun{
b <- "s3://my-crazy-bucket/dir1"
list_s3_objects(b)
}
}
