% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{manta_isec_stats}
\alias{manta_isec_stats}
\title{Manta Comparison Metrics}
\usage{
manta_isec_stats(mi, samplename, flab)
}
\arguments{
\item{mi}{Output from \code{\link{manta_isec}}.}

\item{samplename}{Sample name (used for labelling).}

\item{flab}{File name (used for labelling).}
}
\value{
A single row tibble with the following columns:
\itemize{
\item sample: samplename
\item flabel: flab
\item run1_count: total variants in first file
\item run2_count: total variants in second file
\item TP: in first and second
\item FP: in first but not second
\item FN: in second but not first
\item Recall: TP / (TP + FN)
\item Precision: TP / (TP + FP)
\item Truth: TP + FN
}
}
\description{
Returns evaluation metrics from comparing two Manta call sets.
}
\examples{
\dontrun{
f1 <- "path/to/run1/manta.vcf.gz"
f2 <- "path/to/run2/manta.vcf.gz"
mi <- manta_isec(f1, f2, "sampleA", "manta1")
manta_isec_stats(mi, "sampleA", "manta1")
}

}
