% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio.R
\name{merge_bcbio_outputs}
\alias{merge_bcbio_outputs}
\title{Gather bcbio filepaths from two final directories into a single tibble}
\usage{
merge_bcbio_outputs(f1, f2)
}
\arguments{
\item{f1}{Path to first \code{final} bcbio directory.}

\item{f2}{Path to second \code{final} bcbio directory.}
}
\value{
A tibble with the following columns:
\itemize{
\item ftype: file type.
\item f1: final1 file path
\item f2: final2 file path
}
}
\description{
Generates a tibble containing absolute paths to files from two bcbio final directories.
}
\examples{
\dontrun{
final1 <- "path/to/bcbio/final1"
final2 <- "path/to/bcbio/final2"
merge_bcbio_outputs(final1, final2)
}
}
