% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{merge_umccrise_outputs}
\alias{merge_umccrise_outputs}
\title{Gather umccrise filepaths from two umccrise final directories into a single tibble}
\usage{
merge_umccrise_outputs(d1, d2, sample)
}
\arguments{
\item{d1}{Path to first <umccrised/sample> directory.}

\item{d2}{Path to second <umccrised/sample> directory.}

\item{sample}{Sample name.}
}
\value{
A tibble with the following columns:
\itemize{
\item sample name
\item variant type (e.g. SNV, SV, CNV)
\item file label (e.g. pcgr, cpsr, purple etc.)
\item run1 file path
\item run2 file path
}
}
\description{
Generates a tibble containing absolute paths to (common) files from two umccrise final directories.
}
\examples{
\dontrun{
um1 <- "path/to/umccrised/sample1"
um2 <- "path/to/umccrised/sample2"
merge_umccrise_outputs(um1, um2)
}
}
