% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_cmp}
\alias{multiqc_cmp}
\title{Compare MultiQC JSONs}
\usage{
multiqc_cmp(x1, x2, out)
}
\arguments{
\item{x1}{Path to MultiQC JSON from run A.}

\item{x2}{Path to MultiQC JSON from run B.}

\item{out}{Path to output results.}
}
\value{
Tibble with MultiQC metrics that differ between the two runs.
}
\description{
Compares MultiQC JSON outputs from two runs.
}
\examples{
\dontrun{
x1 <- "multiqc_data1.json"
x2 <- "multiqc_data2.json"
lab1 <- "v212"
lab2 <- "v200"
cmp(x1, x2, lab1, lab2)
}
}
