% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_compare}
\alias{multiqc_compare}
\title{Compare MultiQC JSONs}
\usage{
multiqc_compare(x1, x2, lab1, lab2)
}
\arguments{
\item{x1}{Path to MultiQC JSON from run A.}

\item{x2}{Path to MultiQC JSON from run B.}

\item{lab1}{Label for run A.}

\item{lab2}{Label for run B.}
}
\value{
Tibble with MultiQC metrics that differ between the two runs.
}
\description{
Compares MultiQC JSON outputs from two runs.
}
