% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgr.R
\name{pcgr_cmp}
\alias{pcgr_cmp}
\title{PCGR TSV Comparison}
\usage{
pcgr_cmp(f1, f2)
}
\arguments{
\item{f1}{Path to file1.}

\item{f2}{Path to file2.}
}
\description{
PCGR TSV Comparison
}
\examples{

\dontrun{
f1 <- "~/projects/woof/woof_compare_reports/nogit/umccrise_2.3.0rc.4/umccrised/SEQC50__SEQC-II_tumour/small_variants/SEQC50__SEQC-II_tumour-somatic.pcgr.snvs_indels.tiers.tsv"
f2 <- "~/projects/woof/woof_compare_reports/nogit/umccrise_2.2.0/umccrised/SEQC50__SEQC-II_tumour/small_variants/SEQC50__SEQC-II_tumour-somatic.pcgr.snvs_indels.tiers.tsv"
pcgr_cmp(f1, f2)
}

}
