% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{qc_sum_cmp}
\alias{qc_sum_cmp}
\title{umccrise QC Summary Comparison}
\usage{
qc_sum_cmp(f1, f2, out)
}
\arguments{
\item{f1}{Path to summary table for run1.}

\item{f2}{Path to summary table for run2.}

\item{out}{Path to write results to.}
}
\description{
f1 <- "/Users/pdiakumis/projects/woof_compare_reports/nogit/umccrise_2.0.0/umccrised/SEQC50__SEQC-II_tumour/cancer_report_tables/SEQC50__SEQC-II_tumour-qc_summary.tsv.gz"
f2 <- "/Users/pdiakumis/projects/woof_compare_reports/nogit/umccrise_2.1.2_rc6/umccrised/SEQC50__SEQC-II_tumour/cancer_report_tables/SEQC50__SEQC-II_tumour-qc_summary.tsv.gz"
}
