% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbio.R
\name{read_bcbio_config}
\alias{read_bcbio_config}
\title{Read bcbio Config}
\usage{
read_bcbio_config(x)
}
\arguments{
\item{x}{Path to \verb{bcbio/config/<datestamp>_project.yaml} file.}
}
\value{
A list with the following elements:
\itemize{
\item 'batch_name', 'tumor_name' and 'normal_name': names of batch and its
tumor/normal samples.
\item 'analysis_type': variant2 or RNA-seq
\item 'genome_build': GRCh37 or hg38
\item 'aligner': bwa, star etc.
}
}
\description{
Generates a list of information about the batch within a bcbio
config YAML file.
}
\examples{
\dontrun{
x <- "path/to/bcbio/config/<datestamp>_project.yaml"
read_bcbio_config(x)
}
}
