% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{read_bcbio_configs}
\alias{read_bcbio_configs}
\title{Read YAML configs from two bcbio runs}
\usage{
read_bcbio_configs(conf1, conf2)
}
\arguments{
\item{conf1}{Path to \code{final/config/<timestamp>_project.yaml} for run1.}

\item{conf2}{Path to \code{final/config/<timestamp>_project.yaml} for run2.}
}
\value{
A list with the following elements:
\itemize{
\item 'batch_name', 'tumor_name', 'normal_name': batch/tumor/normal names.
\item 'variant_files': file type. Can be one of:
\enumerate{
\item ensemble_som
\item mutect2_som
\item strelka2_som
\item vardict_som
\item ensemble_ger
\item gatk-haplotype_ger
\item strelka2_ger
\item vardict_ger
\item manta_som-sv
}
}
}
\description{
Generates a list containing absolute paths to variant files for
two bcbio runs.
}
\examples{
\dontrun{
conf1 <- "path/to/bcbio_run1/final/config/<timestamp>_project.yaml"
conf2 <- "path/to/bcbio_run2/final/config/<timestamp>_project.yaml"
read_bcbio_configs(conf1, conf2)
}
}
