% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{read_bcftools_querysn}
\alias{read_bcftools_querysn}
\title{Read bcftools query output}
\usage{
read_bcftools_querysn(f)
}
\arguments{
\item{f}{Path to file output by bcftools query.}
}
\value{
Tibble with three columns: validation verdict, input file name, and basename.
}
\description{
Reads output from bcftools query for sample list.
}
