% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{read_count_file}
\alias{read_count_file}
\title{Read SNV count file output by woof}
\usage{
read_count_file(x)
}
\arguments{
\item{x}{Path to file with counts}
}
\value{
A single integer
}
\description{
Reads SNV count file output by woof
}
\examples{

x <- system.file("extdata", "count_vars.txt", package = "woofr")
read_count_file(x)

}
