% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{read_fqtools}
\alias{read_fqtools}
\title{Read fqtools output}
\usage{
read_fqtools(f)
}
\arguments{
\item{f}{Path to file output by fqtools.}
}
\value{
Tibble with two columns: validation verdict and input file name.
}
\description{
Reads output from fqtools.
}
