% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{read_md5sum}
\alias{read_md5sum}
\title{Read md5sum output}
\usage{
read_md5sum(f)
}
\arguments{
\item{f}{Path to md5sum file.}
}
\value{
Tibble with two columns: md5sum and basename of input file.
}
\description{
Reads output md5sum.
}
