% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{read_purple_gene_file}
\alias{read_purple_gene_file}
\title{Read PURPLE gene segments}
\usage{
read_purple_gene_file(x)
}
\arguments{
\item{x}{Path to PURPLE \code{gene.cnv} (or \code{cnv.gene.tsv}) file.}
}
\value{
A tibble with the main columns of interest from the PURPLE \code{gene.cnv} file i.e.
chrom, start, end, gene, min_cn and max_cn.
}
\description{
Reads the gene.cnv file output by PURPLE
}
\examples{

x <- system.file("extdata", "cnv/sample_A.purple.gene.cnv", package = "woofr")
y <- system.file("extdata", "cnv/sample_B.purple.cnv.gene.tsv", package = "woofr")
read_purple_gene_file(x)
read_purple_gene_file(y)

}
