% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{read_qcheck}
\alias{read_qcheck}
\title{Read samtools quickcheck output}
\usage{
read_qcheck(f)
}
\arguments{
\item{f}{Path to file output by samtools quickcheck.}
}
\value{
Tibble with two columns: validation verdict and input file name.
}
\description{
Reads output from samtools quickcheck.
}
