% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{read_snv_eval_file}
\alias{read_snv_eval_file}
\title{Read SNV eval file output by woof}
\usage{
read_snv_eval_file(x)
}
\arguments{
\item{x}{Path to file with eval results}
}
\value{
A tibble with a single row of evaluation metrics
}
\description{
Reads SNV eval file output by woof
}
\examples{

x <- system.file("extdata", "snv/eval_stats.tsv", package = "woofr")
read_snv_eval_file(x)

}
