% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{read_sv_eval_file}
\alias{read_sv_eval_file}
\title{Read SV eval metrics file output by woof}
\usage{
read_sv_eval_file(x)
}
\arguments{
\item{x}{Path to file with SV evaluation metrics}
}
\value{
A tibble with a single row of evaluation metrics
}
\description{
Reads SV eval metrics file output by woof
}
\examples{

x <- system.file("extdata", "sv/eval_metrics.tsv", package = "woofr")
read_sv_eval_file(x)

}
