% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{read_sv_fpfn_file}
\alias{read_sv_fpfn_file}
\title{Read SV FP/FN file output by woof}
\usage{
read_sv_fpfn_file(x)
}
\arguments{
\item{x}{Path to file with FP/FN variants}
}
\value{
A tibble containing FP/FN SV variants
}
\description{
Reads SV FP/FN file output by woof
}
\examples{

x <- system.file("extdata", "sv/fpfn.tsv", package = "woofr")
read_sv_fpfn_file(x)

}
