% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{stock}
\alias{stock}
\alias{stocks}
\title{Scraps the historical stock prices for the provided companies}
\usage{
stock(x)

stocks(...)
}
\arguments{
\item{x, ...}{Company name as returned by the \code{\link[=get_today]{get_today()}} function}
}
\description{
Scraps \href{http://leboursier.ma/}{\strong{Le Boursier}}'s API and get the historical prices for the specified companies.

Returns a data.frame containing all the historical data for the specified stock.
If the plural variant is used it'll add a \code{symbol} column that'll contain the name of the company.
}
\examples{
\dontrun{
  stock("Wafa Assur")
  stock() # same as runing stock("MASI")
  stock("MADEX")}
\dontrun{
  stocks("Wafa Assur", "Nexans Maroc")
  stocks("Masi", "MADEX")}
}
