% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree}
\alias{tree}
\title{Call system tree}
\usage{
tree(x, ...)
}
\arguments{
\item{x}{File path.}

\item{...}{Passed to \code{paste0}. Stuff to put after \code{tree_}.}
}
\value{
A \code{tree} representation of \code{x}.
}
\description{
Calls the system tree.
}
