% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{umccrise_outputs}
\alias{umccrise_outputs}
\title{Get umccrise final output files}
\usage{
umccrise_outputs(d)
}
\arguments{
\item{d}{Path to \verb{<umccrised/sample>} umccrise directory.}
}
\value{
A tibble with the following columns:
\itemize{
\item vartype: variant type. Can be one of:
\itemize{
\item SNV (single-nucleotide/indel variants)
\item SV (structural variants)
\item CNV (copy number variants)
}
\item flabel: file label (e.g. pcgr, cpsr, purple etc.)
\item fpath: path to file
}
}
\description{
Generates a tibble containing absolute paths to umccrise final files.
}
\examples{
\dontrun{
um <- "path/to/umccrised/sample"
umccrise_outputs(um)
}
}
