% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{bcbio_outputs}
\alias{bcbio_outputs}
\title{Get bcbio final output files}
\usage{
bcbio_outputs(d)
}
\arguments{
\item{d}{Path to \code{<bcbio/final>} directory.}
}
\value{
A tibble with the following columns:
\itemize{
\item vartype: variant type. Can be one of:
\itemize{
\item snv (single-nucleotide/indel variants)
\item sv (structural variants)
}
\item flabel: file label (e.g. ensemble, manta, vardict etc.)
\item fpath: path to file
}
}
\description{
Generates a tibble containing absolute paths to bcbio final files.
}
\examples{
\dontrun{
d <- "path/to/bcbio/final"
bcbio_outputs(d)
}
}
