% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonds.R
\name{bonds}
\alias{bonds}
\title{Scraps the data of the Moroccan Bonds}
\usage{
bonds()
}
\description{
Returns a data.frame containing all the bonds listed in \href{http://leboursier.ma/details/marche.html}{\strong{Le Boursier}}.
The column names are self explanatory except for \code{full_name} which stands for the full name of the bond issuer.
}
\examples{
\dontrun{
  bonds()}
}
