% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{get_circos}
\alias{get_circos}
\title{Get Manta Circos with FP/FN calls}
\usage{
get_circos(mi, samplename, outdir)
}
\arguments{
\item{mi}{Output from \code{\link{manta_isec}}.}

\item{samplename}{Sample name (used for labelling).}

\item{outdir}{Directory to write circos to.}
}
\value{
Path to circos plot highlighting FP (green) and FN (red) calls from a
Manta comparison.
}
\description{
Returns Circos plot highlighting FP/FN calls from a Manta comparison.
}
\examples{

\dontrun{
f1 <- "path/to/run1/manta.vcf.gz"
f2 <- "path/to/run2/manta.vcf.gz"
mi <- manta_isec(f1, f2)
get_circos(mi, "sampleA", "outdir1")
}

}
