% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_today.R
\name{get_today}
\alias{get_today}
\title{Get the daily variation}
\usage{
get_today()
}
\description{
Returns a data.frame containing the daily variation for all the stocks
}
\examples{
\dontrun{
  get_today()}
}
