% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\alias{sector_indices}
\title{Moroccan Market and Sector indices}
\usage{
indices()

sector_indices()
}
\description{
Gets the Moroccan Market and Sector indices that are listed in \href{http://leboursier.ma/details/marche.html}{\strong{Le Boursier}}.
}
\examples{
  indices()
  sector_indices()
}
