% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\alias{sector_indices}
\title{Gets the Moroccan Market and Sector indices}
\usage{
indices()

sector_indices()
}
\description{
\code{indices} (resp. \code{sector_indices}) Returns a data.frame containing all the Market Indices (resp. Sector Indices) listed in \href{http://leboursier.ma/details/marche.html}{\strong{Le Boursier}}.
}
\details{
The column names are self explanatory.
}
\examples{
\dontrun{
  indices()}
\dontrun{
  sector_indices()}
}
