% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_day.R
\name{intraday}
\alias{intraday}
\alias{intradays}
\title{Intraday Prices}
\usage{
intraday(name)

intradays(...)
}
\arguments{
\item{name, ...}{Company names as returned by the \code{\link[=get_today]{get_today()}} function}
}
\value{
Returns a \code{data.frame} containing all the intraday prices for the specified stocks.
\itemize{
\item \code{time}: the time of the record.
\item \code{price}: the price at that time.
Where the \code{symbol} column contains the name of the company.
}
}
\description{
Scraps the intraday stock prices for the provided companies
}
\examples{
head(intraday(), 30)
intradays("Wafa Assur", "Nexans Maroc")
}
