% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_day.R
\name{intraday}
\alias{intraday}
\alias{intradays}
\title{Intraday Prices}
\usage{
intraday(name)

intradays(...)
}
\arguments{
\item{name, ...}{Company names as returned by the \code{\link[=get_today]{get_today()}} function}
}
\description{
Intraday Prices
}
\details{
Scraps the intraday stock prices for the provided companies

Returns a data.frame containing all the intraday prices for the specified stocks.
Where the \code{symbol} column contains the name of the company.
}
\examples{
  intraday("Wafa Assur")
  intradays("Wafa Assur", "Nexans Maroc")
}
