% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{merge_umccrise_outputs}
\alias{merge_umccrise_outputs}
\title{Gather umccrise filepaths from two different sample directories into a single tibble}
\usage{
merge_umccrise_outputs(d1, d2)
}
\arguments{
\item{d1}{Path to first umccrise sample directory.}

\item{d2}{Path to second umccrise sample directory.}
}
\value{
A tibble with the following columns:
\itemize{
\item ftype: file type.
\item d1: sample1 file path
\item d2: sample2 file path
}
}
\description{
Generates a tibble containing absolute paths to files from two umccrise sample directories.
}
\examples{
\dontrun{
um1 <- "path/to/umccrised/sample1"
um2 <- "path/to/umccrised/sample2"
merge_umccrise_outputs(um1, um2)
}
}
