% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{umccrise_outputs}
\alias{umccrise_outputs}
\title{Get umccrise final output files}
\usage{
umccrise_outputs(d)
}
\arguments{
\item{d}{Path to \verb{<umccrised/sampleid__tumorid>} umccrise directory.}
}
\value{
A tibble with the following columns:
\itemize{
\item ftype: file type. Can be one of:
\itemize{
\item snv (single-nucleotide/indel variants VCF)
\item sv (Manta structural variants VCF)
\item cnv (PURPLE copy number variants)
\item multiqc (MultiQC JSON)
\item hrd (HRDetect or CHORD TSV results)
\item canrep_qcsum (QC summary from Cancer Report Rmd)
}
\item flabel: file label (e.g. multiqc, sage, purple etc.)
\item fpath: path to file
}
}
\description{
Generates a tibble containing absolute paths to umccrise final files.
}
\examples{
\dontrun{
um <- "path/to/umccrised/sample"
umccrise_outputs(um)
}
}
