% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{umccrise_outputs}
\alias{umccrise_outputs}
\title{Get final umccrise sample files}
\usage{
umccrise_outputs(d)
}
\arguments{
\item{d}{Path to \code{umccrised/<batch>__<tumor_name>} umccrise directory.}
}
\value{
A tibble with the following columns:
\itemize{
\item fpath: path to file
\item ftype: file type. Can be one of:
\enumerate{
\item pcgr
\item cpsr
}
}
}
\description{
Generates a tibble containing absolute paths to final umccrise sample files.
}
\examples{
\dontrun{
um <- "path/to/umccrised/sample"
umccrise_outputs(um)
}
}
