% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woofr-package.R
\docType{package}
\name{woofr-package}
\alias{woofr}
\alias{woofr-package}
\title{woofr}
\description{
Helper functions for woof.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pdiakumis/woofr}
  \item Report bugs at \url{https://github.com/pdiakumis/woofr/issues}
}

}
\author{
\strong{Maintainer}: Peter Diakumis \email{peterdiakumis@gmail.com} (0000-0002-7502-7545)

}
